updated scoring extension
Now it's using api v3, that changes the URL so I also had to create new
configuration key for URL to maintain compatibility with older versions
of this extension.

We are using "damaging" model as default and fallback to reverted in
case it's not supported by ORES. Goodfaith score is also collected, but
not really used right now (except it's being displayed in metadata).